- .NET Code was cleared and brought to a readable form. - Added an some custom exception logger
I didn't change the JS code, cuz it's a prototype of future service and frontend of this will be created again (may be using Razor Pages).

Also this app has specific architecture for ASP.NET 6 Core initialy and I want to create new app based on this in future with right creating sequence.